Support subsetting rows within a table.
This change adds 3 main pieces of functionality:
 1. Adds `PTable[PIndexSet]` to allow gathering a subset of rows from within a table.
 2. Adds `PTable.count` to easily determine the number of rows within a table.
 3. Conforms `PTable` to `Equatable` to allow easier comparison of two tables.

Note: in the process, I've defined a "poor man's" `==` for `PColumn`, in order to
work around limitations of Swift's existential implementations. (i.e. If `PColumn`
were to refine `Equatable`, because `Equatable` uses `Self`, it could no longer be
used as an existential type, and thus could not be referenced as a standalone type.)

Encountered [SR-11748](https://bugs.swift.org/browse/SR-11748) in the process.