Add better support for optionals.
After playing around with a few different representations for handling optionals,
I settled on this design. Although this will require heavily optimizing
`PIndexSet`s, I believe this is ergonomic and effective. This design allows
any datatype to also contain `nil`s. Additionally, it _radically_ simplifies
the internals, as we can now assume `T`s are `Equatable`, `Comparable`, and
`Hashable`, and are `DefaultInit`-able.

As part of putting together this patch, I played around with a variety of
representations for the backing data for `PIndexSet`. Supporting automatically
extending the shorter one makes things tricky. It might make sense to stop
supporting this functionality, or to make it explicit.

Finally, this representation will be fully amenable to parallelization in
the future when the internals are parallelized, however it will require
support for parallel_zip as a combinator of parallel iterators.