Fix race condition in TransformPipelineIterator.
Note: I'm also checking in the test program I wrote to flush out
the errors, as it's much easier to determine the issues when running
under lldb, which is much easier when you have a real program to
debug. The following commit will remove it. (This is just to preserve
it in history.)