Implement pmap on Array.
In the process, I debugged actual parallel execution. Note that for
short-term correctness, I've added a bunch of extra locks. This is because
the focus is not on performance in the short term, but rather to ensure
correctness.

Note: I'm also checking in the testing executable that I used to test with
lldb. This will be deleted in a subsequent commit.