Optimize the performance of CSV Parsing.
Profiling the previous implementation revealed that over 92% of the
time spent in parseRow was spent in String(bytesNoCopy:...), as it
was spending all its time validating the utf8 format. Additionally, over
93% of the time spend adding the cells to the typedcolumn's is spent in
foundation performing character trimming (i.e. parsing from ascii).

This change refactors to avoid using String's when convenient. It adds
a CSVCell enum which differentiates between empty cells, a raw buffer
reference, and a parsed (e.g. due to escaping) cell that's a string
(this is a slow path).

Additionally, to make this ergonomic, this change defines a PCSVParsible
protocol, and provides a default implementation. It also adds fast
implementations for `Int` and `Double` (the two expected most common
and optimizable types in CSV parsing).