Refactor `GraphVisitor` & family, makes `TablePredecessorVisitor` work for all searches.
This change consolidates all the visitors into a single file, and adds a few new protocols
into the hierarchy to reduce duplication of methods. By doing so, it is now possible to
easily and simply conform a single visitor to work across multiple graph searches.

In order to simplify the protocol hierarchy, I've modified the contract for DFS & BFS to
both call `start(vertex:_:)`.