Add additional assertion to ensure the Swift compiler isn't too clever.
In optimized builds of the test suite, the Swift compiler is smart enough to deallocate
`b` before we can assert that not all `MemoryChecker()` instances have been cleaned up.
This change adds an additional assertion on `b` after we can verify this property, which
guarantees that `b` lives long enough (and thus the instance of the `MemoryChecker()`)
to perform a proper test.