Allow arbitrary output sinks for topological sort.
This change makes the topological sort implementation more generic
by allowing an arbitrary way of consuming the output of the topological
sort algorithm. In reverse topological order, we call `reverseSink`, a
function of type `(Graph.VertexId) -> Void`.

This change then reimplements the original topological sort in terms of
the this newly refactored `topologicalSort` implementation.