Move `startingAt` to be first argument of DFS and add overload.
This commit moves `start` (which was the final argument to `depthFirstSearch`
to `startingAt`, and moving it to be the first argument. Additionally, this
change adds an overload of `depthFirstSearch` that doesn't require a
`VertexVisitationState` (previously `ColorMap`), to make it a nicer API.
Finally, it updates one of the tests to use this new overload.