Simplify Property Map's
Previously, things were separated into two parallel protocol families of
vertex property maps, and edge property maps. This made re-use of property
map implementations more difficult, as they had to be implemented for both
vertices as well as edges. Further, there was a non-trivial amount of
syntactic overhead for external property maps.

This change simplifies property maps by adding an extra associated type
to the base protocol. Once that has been added, it becomes trivial to
make an implementation that is generic over VertexId's and EdgeId's.
This has resulted in a significant simplification.

In addition, I've (for now) consolidated mutable and immutable property
map protocols into a single protocol. (This design decision might be
good to revisit in the future if this turns out to be problematic. For
the existing algorithms, this is not.)

There are a few bits of follow-up:

 1. Determine if it's possible to make helpful typealiases to reduce
    the amount of typing required to specify the where clause for a
    property map. Something like the following (which unfortunately
    doesn't compile):

    ```swift
    extension GraphProtocol {
      typealias VertexPropertyMap = PropertyMap where Graph == Self, Key == VertexId
    }
    ```

 2. Explore the space of property maps for use in vertex-parallel
    algorithms.
 3. Reevaluate the name `PropertyMap`.