Make AdjacencyList actually execute in parallel.
Previously, `step` actually called `sequentialStep` instead of `parallelStep`.
This, unsurprisingly, did not yield parallelism. In order to make this change,
it required defining a "parallel projection" of the graph. This is because
graphs commonly implement value semantics. Unfortunately, value semantics is
somewhat incompatible with simultaneous mutation by multiple threads. A
parallel projection is a transformation into the parallel domain such that it
can implement buffer semantics.