Make AdjacencyList (almost) actually execute in parallel. (#19)
* Make AdjacencyList actually execute in parallel.

Previously, `step` actually called `sequentialStep` instead of `parallelStep`.
This, unsurprisingly, did not yield parallelism. In order to make this change,
it required defining a "parallel projection" of the graph. This is because
graphs commonly implement value semantics. Unfortunately, value semantics is
somewhat incompatible with simultaneous mutation by multiple threads. A
parallel projection is a transformation into the parallel domain such that it
can implement buffer semantics.

* Reset `hasMessages` flag after delivering in PerThreadMailboxes.

Previously, we would sometimes have an extra round of delivering if the
previous step had messages sent from multiple threads. This change writes
a test that reproduces the failure, and also includes the fix.

* Add an extra benchmark to demonstrate performance improvements.

* Revert change requiring external property maps for edge properties.

Note: this change doesn't actually make things execute in parallel; overheads are still higher than I'd like. So for now, we just execute serially.