Refactor to use POD type in TaskDeque.
This change refactors `NonBlockingThreadPool` to use a POD-type instead of
moving around closures all the time. This change by itself does not improve
performance, however it does set up further optimizations.

As part of this change, `TaskDeque` has been refactored to return `true`
or `false` depending on whether the item was enqueued or not. This is
easier to reason about; since the data we're storing is not `moveonly` but
is instead a POD type, this does not inhibit performance.