[ThreadPool] Special-case parallelFor in work structure.
By making this change, we set ourselves up for being able to remove the
ARC traffic in the next commit. This change by itself does not speed up
the thread pool parallel for implementation at one level, but does yield
a ~1.5x speed-up when nesting parallelism (500 us -> 350 us in the
benchmark).