Rewrite the heap and related priority queue structures.
This change moves the heap related algorithms to operate generically on
`Collection` (and its various refinements). In doing so, the algorithms
become far more general and reusable.

Further, we re-build the PriorityQueue type, solving the indexing TODO's
that plagued the previous implementation.

Finally, we update all the rest of the code to use the new API.