Bidirectional & Undirected AdjacencyLists (#50)
This change (1) implements bidirectional and undirected adjacency lists (to complement the existing directed implementation, and (2) defines a variety of protocols & types in order to attempt to unify AdjacencyList implementations. Unfortunately, this triggered a variety of mainline Swift compiler crashers (that mostly don't seem to affect S4TF toolchains). As a result, there's far more duplication than would be possible in the future.