Support easy copying from one graph representation to another.
This change adds support for initializing a new mutable graph from an
older graph by crawling over the graph and converting it. In addition,
this changes the default types in the SimpleAdjacencyLists to be `Int`'s
because that's definitely simpler to use. Further, it adds support for
SearchDefaultsGraph to all AdjacencyLists and to the EdgeFilterGraph.