SourceInitializableCollection and withContiguous{Mutable}StorageIfAvailable for FixedSizeArray (#102)
In https://github.com/borglab/SwiftFusion/pull/148, we need to initialize a collection of known type (a collection of jacobian matrix rows) from a lazy map. The known type can be a `FixedSizeArray` or an `Array`. This PR provides a `SourceInitializableCollection` protocol with an initializer that does this.

We also need to access contiguous storage in a way that generalizes across `FixedSizeArray` and `Array`, so this PR implements `withContiguous{Mutable}StorageIfAvailable` for `FixedSizeArray`.